% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_area.R
\name{convert_area}
\alias{convert_area}
\title{Convert between area units--miles2, ft2, inch2, cm2, m2, km2}
\usage{
convert_area(x, from, to)
}
\arguments{
\item{x}{A vector of numbers to be converted.}

\item{from}{A character vector of the units x is in. Must be length 1 or same length as x.}

\item{to}{A character vector of the units into which x will be converted. Must be length 1 or same length as x.}
}
\value{
A vector of numbers converted FROM old units TO new units.
}
\description{
Convert between area units--miles2, ft2, inch2, cm2, m2, km2
}
\examples{
convert_area(1, "ft2", "inch2")
convert_area(c(1, 10, 100), "ft2", "inch2")
convert_area(c(1, 1, 1), c("ft2", "inch2", "cm2"), c("inch2", "cm2", "ft2"))
}
