% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_density.R
\name{convert_density}
\alias{convert_density}
\title{Convert between density units -- kg/m3, g/L, mg/L, lb/gal, lb/ft3}
\usage{
convert_density(x, from, to)
}
\arguments{
\item{x}{A vector of numbers to be converted.}

\item{from}{A character vector of the units x is in. Must be length 1 or same length as x.}

\item{to}{A character vector of the units into which x will be converted. Must be length 1 or same length as x.}
}
\value{
A vector of numbers converted FROM old units TO new units.
}
\description{
Convert between density units -- kg/m3, g/L, mg/L, lb/gal, lb/ft3
}
\examples{
convert_density(1, "kg/m3", "lb/ft3")
convert_density(c(1, 10, 100), "kg/m3", "lb/ft3")
convert_density(c(1, 10, 100), c("g/l", "lb/gal", "lb/ft3"), c("lb/ft3", "g/l", "lb/gal"))
}
