% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Q.R
\name{convert_Q}
\alias{convert_Q}
\title{Convert between flow units typically seen in wastewater and streamflow calculations--cfs, acre-ft/d, mgd, lpm, cms}
\usage{
convert_Q(x, from, to)
}
\arguments{
\item{x}{A vector of numbers to be converted.}

\item{from}{A character vector of the units x is in. Must be length 1 or same length as x.}

\item{to}{A character vector of the units into which x will be converted. Must be length 1 or same length as x.}
}
\value{
A vector of numbers converted FROM old units TO new units.
}
\description{
Convert between flow units typically seen in wastewater and streamflow calculations--cfs, acre-ft/d, mgd, lpm, cms
}
