% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{convert_temp}
\alias{convert_temp}
\title{Convert temperature units--C, F, K, and R}
\usage{
convert_temp(x, from, to)
}
\arguments{
\item{x}{A vector of numbers to be converted.}

\item{from}{A character vector of the units x is in. Must be length 1 or same length as x.}

\item{to}{A character vector of the units into which x will be converted. Must be length 1 or same length as x.}
}
\value{
A vector of numbers converted FROM old units TO new units.
}
\description{
Convert temperature units--C, F, K, and R
}
